Move playground routes under `/stuff`
