--- src/app/(playground)/inm-clock/meta.ts
@@ -1,30 +1,0 @@
-import type { Meta } from "@/types/meta";
-
-export const meta: Meta = {
-	title: "迫真時鐘部．時停の裏技",
-	description: "這麼惡臭的時鐘有存在的必要嗎（惱",
-	url: "/tools/inm-clock",
-	keywords: [
-		// Main Keywords
-		"inm clock",
-		"時鐘工具",
-		"web clock widget",
-		"網頁時鐘元件",
-		"2025 時鐘",
-		"114 時鐘",
-		"114年 時鐘",
-
-		// Kuso Keywords
-		"inm",
-		"114514",
-		"淫夢",
-		"野獸先輩",
-		"野獣先輩",
-		"仲夏夜之淫夢",
-		"真夏の夜の淫夢",
-		"田所浩二",
-		"こ↑こ↓",
-		"homo",
-		"ホモ",
-	],
-};
