+++ src/app/stuff/inm-clock/icon.png
